chore: refactor file system (uploads)
