+++ .env.example
@@ -55,7 +55,7 @@ MONGO_PASSWORD=
 
 LOG_MONGO_COLLECTION=
 
-CLOUD_STORAGE=s3
+FILESYSTEM_DRIVER=gcs
 
 AWS_ACCESS_KEY_ID=
 AWS_SECRET_ACCESS_KEY=
