+++ src/config/index.ts
@@ -39,4 +39,6 @@ config.set('gcloud.key', decodeBase64(config.get('gcloud.key')))
 config.set('file.type', config.get('file.type', 'jpg|png|svg|jpeg|pdf'))
 config.set('file.max', config.get('file.max', 10)) // Unit MB
 
+config.set('filesystem.driver', config.get('filesystem.driver', 'gcs'))
+
 export default config
