+++ src/emails/invitation_partner.handlebars
@@ -19,7 +19,7 @@
   <body style="background-color: #efefef; margin: 0;" class="body-night-scheme">
     <div style="margin: 16px auto; padding: 0px 16px; max-width: 600px;">
       <section class="card-night-scheme" style="margin-bottom: 16px; background-color: #ffffff; border-radius: 16px; height: auto;">
-        <img style="width: 100%; max-width: 600px; height: auto; border-radius: 16px 16px 0px 0px;" width="382px" height="110px" src="https://dedi-cdn.digitalservice.id/assets/header_undang_dekstop.png" alt="greeting person">
+        <img style="width: 100%; max-width: 600px; height: auto; border-radius: 16px 16px 0px 0px;" width="382px" height="110px" src="{{ storageUrl }}/assets/header_undang_dekstop.png" alt="greeting person">
         <div style="padding: 16px; text-align: center;">
           <h1 style="color: #022B55; font-family: 'Lato', sans-serif; font-size: 14px; line-height: 23px; font-weight: 700;">
               Hai, {{ name }}!
@@ -38,7 +38,7 @@
               Dui viverra quam lacus, proin ac pulvinar quis nec. Vel dictumst.
           </p>
           <div class="info-night-scheme" style="padding: 10px 16px; display: inline-flex; column-gap: 10px; background-color: #E6F7FF; border: 1px solid #53bdff; border-radius: 8px;">
-            <img width="17" height="17" src="https://dedi-cdn.digitalservice.id/assets/icon_ask.png" alt="icon_ask">
+            <img width="17" height="17" src="{{ storageUrl }}/assets/icon_ask.png" alt="icon_ask">
             <p style="text-align: left; margin: 0; color: #022B55; font-family: 'Lato', sans-serif; font-size: 14px; line-height: 23px; font-weight: 400;">
               Apabila ada pertanyaan atau kamu membutuhkan bantuan,
               silahkan menghubungi kami melalui <span style="text-decoration: underline; font-weight: 700"><a style="color: #1e88e5;">Help Desk</a></span>
@@ -47,7 +47,7 @@
         </div>
       </section>
       <footer style="text-align: left; width: 100%;">
-        <img width="158" height="32" src="https://dedi-cdn.digitalservice.id/assets/logo_full_desa_digital.png" alt="image_text_desa_digital">
+        <img width="158" height="32" src="{{ storageUrl }}/assets/logo_full_desa_digital.png" alt="image_text_desa_digital">
         <p style="color: #929292; font-family: 'Lato', sans-serif; font-size: 12px; line-height: 20px; font-weight: 600;">
           Copyright &copy; 2022
         </p>
