+++ src/helpers/mail.ts
@@ -3,6 +3,7 @@ import { TemplateOptions } from 'nodemailer-express-handlebars'
 import { Options } from 'nodemailer/lib/mailer'
 import config from '../config'
 import mail from '../config/mailer'
+import { StorageUrl } from './storage'
 
 export type Payload = Options & TemplateOptions
 
@@ -18,4 +19,8 @@ export const sendMail = async (payload: Payload) => {
   mail.sendMail(mailOptions, (error: Error) => {
     if (error) newrelic.noticeError(error)
   })
+}
+
+export const contextDefault = {
+  storageUrl: StorageUrl(),
 }
