+++ src/helpers/uploads/gcs.ts
@@ -1,14 +1,10 @@
 import { Express } from 'express'
-import { v4 as uuidv4 } from 'uuid'
-import path from 'path'
 import newrelic from 'newrelic'
-import config from '../config'
-import { GCS } from '../config/cloudStorage'
+import config from '../../config'
+import { GCS } from '../../config/cloudStorage'
 
-export const uploadGCS = (file: Express.Multer.File): string => {
+export const uploadGCS = (file: Express.Multer.File, Key: string) => {
   const bucket = GCS.bucket(config.get('gcs.bucket'))
-  const filename = uuidv4() + path.extname(file.originalname)
-  const Key = `${config.get('node.env')}/${filename}`
 
   const blob = bucket.file(Key)
   const blobStream = blob.createWriteStream({
@@ -20,6 +16,4 @@ export const uploadGCS = (file: Express.Multer.File): string => {
   })
 
   blobStream.end(file.buffer)
-
-  return filename
 }
